feat: update scrcpy-server 3.3.1

+++ src/device/controller/bufferutil.cpp
@@ -14,7 +14,7 @@ void BufferUtil::write64(QBuffer &buffer, quint64 value)
     write32(buffer, (quint32)value);
 }
 
-void BufferUtil::write16(QBuffer &buffer, quint32 value)
+void BufferUtil::write16(QBuffer &buffer, quint16 value)
 {
     buffer.putChar(value >> 8);
     buffer.putChar(value);
